feat. constants 분리
