+++ lib/rules/object-sort.js
@@ -15,6 +15,13 @@ const {
   useHasGroupComment,
 } = require("../utils");
 
+const {
+  DEFAULT_FUNCTION_KEY_NAMES,
+  DEFAULT_OBJECT_SORTED_ORDER,
+  DEFAULT_FUNCTION_VALUE_TYPES,
+  DEFAULT_FUNCTION_KEY_NAME_PATTERNS,
+} = require("../constants");
+
 //------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
@@ -39,35 +46,24 @@ module.exports = {
               enum: compareKeys,
               type: "string",
             },
-            default: [
-              "isFunctionAsc",
-              "isShortAsc",
-              "keyNameLengthAsc",
-              "localeCompareAsc",
-            ],
+            default: DEFAULT_OBJECT_SORTED_ORDER,
             uniqueItems: true,
           },
           functionKeyNames: {
             type: "array",
             items: { type: "string" },
-            default: ["select", "next"],
+            default: DEFAULT_FUNCTION_KEY_NAMES,
           },
           functionValueTypes: {
             type: "array",
             items: { type: "string" },
-            default: [
-              "AsyncFunction",
-              "MethodDefinition",
-              "GeneratorFunction",
-              "FunctionExpression",
-              "ArrowFunctionExpression",
-            ],
+            default: DEFAULT_FUNCTION_VALUE_TYPES,
           },
           ignoreCommentGroup: { type: "boolean", default: false },
           functionKeyNamePatterns: {
             type: "array",
             items: { type: "string" },
-            default: ["^on", "^callback"],
+            default: DEFAULT_FUNCTION_KEY_NAME_PATTERNS,
           },
         },
         additionalProperties: false,
