+++ lib/utils.js
@@ -1,28 +1,12 @@
-const DEFAULT_OBJECT_SORTED_ORDER = [
-  "isFunctionAsc",
-  "isShortAsc",
-  "keyNameLengthAsc",
-  "localeCompareAsc",
-];
+const {
+  DEFAULT_JSX_SORTED_ORDER,
+  DEFAULT_FUNCTION_KEY_NAMES,
+  DEFAULT_OBJECT_SORTED_ORDER,
+  DEFAULT_FUNCTION_VALUE_TYPES,
+  DEFAULT_FUNCTION_KEY_NAME_PATTERNS,
+} = require("./constants");
 
-const DEFAULT_JSX_SORTED_ORDER = [
-  "isShortAsc",
-  "isFunctionAsc",
-  "keyNameLengthAsc",
-  "localeCompareAsc",
-];
-
-const DEFAULT_FUNCTION_KEY_NAMES = ["select", "next"];
-const DEFAULT_FUNCTION_VALUE_TYPES = [
-  "AsyncFunction",
-  "MethodDefinition",
-  "GeneratorFunction",
-  "FunctionExpression",
-  "ArrowFunctionExpression",
-];
-const DEFAULT_FUNCTION_KEY_NAME_PATTERNS = ["^on", "^callback"];
-
-const COMPARE_MAP = {
+const compareMap = {
   default: () => 0,
   isShortAsc: (a, b) => {
     return a.isShort && !b.isShort ? -1 : !a.isShort && b.isShort ? 1 : 0;
@@ -63,7 +47,7 @@ const useCompare = ({ sortedOrders } = {}, jsx = false) => {
       : DEFAULT_OBJECT_SORTED_ORDER;
   return (a, b) => {
     for (let i = 0; i < orders.length; i++) {
-      const select = COMPARE_MAP[orders[i]] ?? COMPARE_MAP.default;
+      const select = compareMap[orders[i]] ?? compareMap.default;
       const result = select(a, b);
       if (result !== 0) {
         return result > 0 ? 1 : -1;
@@ -168,5 +152,5 @@ module.exports = {
   useCheckFunction,
   useReplaceByNodes,
   useHasGroupComment,
-  compareKeys: Object.keys(COMPARE_MAP),
+  compareKeys: Object.keys(compareMap),
 };
